can control motors, timeout added to recv, etc.
Motors can be controlled, but simulation doesn't seem to always be sending
the correct data. Need to check for that.

A timeout was added to recv in this case that no data is sent because
simulation dropped. This way the last command doesn't keep driving the
motors. Instead, if a timeout has been detected, the minibot stops.

There's a problem with the velocities being given to the minibot.